feat: implement TagCategoryViewController
#14
